requirements.txt
discord
discord.py
python-telegram-bot
python-dotenv